ajouté les verif

+++ modif.php
@@ -1,12 +1,15 @@
 <?php
+session_start();
+  if(!empty($_SESSION['user'])){
      if (isset($_GET['idCommande']) && isset($_GET['delete'])) // Si idCommande et delete sont paramétré 
        {
        		 $db = new PDO('mysql:host=localhost;port=3308;dbname=testcommande','root',''); //On se connecte a la BDD
-           $reponse = $db->prepare('DELETE FROM commande WHERE idCommande= ?'); //Premiére requête pour supprimer l'idCommande 
-       		 $reponse1 = $db->prepare('DELETE FROM produitscommande WHERE Commande_idCommande= ?');
-       		  echo $reponse->execute(array($_GET['idCommande']));
-
+           $reponse = $db->prepare('DELETE FROM commande WHERE idCommande= ? AND Utilisateur_idUser = ?'); //Premiére requête pour supprimer l'idCommande 
+            $reponse1 = $db->prepare('DELETE FROM produitscommande WHERE Commande_idCommande= ?');
+            
+       		 $reponse->execute(array($_GET['idCommande'], $_SESSION['user']));
             $reponse1->execute(array($_GET['idCommande']));
+
             $done[] = true ;
             header('Content-Type: application/json');  // On renvoie le code en JSON 
             echo json_encode($done, JSON_PRETTY_PRINT);
@@ -23,6 +26,11 @@
         header('Content-Type: application/json');  // On renvoie le code en JSON 
         echo json_encode($done, JSON_PRETTY_PRINT);
      } 
-     
+  }
+  else {
+    $wu[] = "Wrong user !";
+    header('Content-Type: application/json');  // On renvoie le code en JSON 
+    echo json_encode($wu, JSON_PRETTY_PRINT);
+  }
      ?>
